fix clippy errors

+++ src/bitmap/mod.rs
@@ -78,7 +78,7 @@ impl PdfiumBitmap {
     /// Returns the height of the image in the bitmap buffer backing this [`PdfiumBitmap`].
     #[inline]
     pub fn height(&self) -> i32 {
-        lib().FPDFBitmap_GetHeight(self) as i32
+        lib().FPDFBitmap_GetHeight(self)
     }
 
     /// Returns the pixel format of the image in the bitmap buffer backing this [`PdfiumBitmap`].
